fix: correct build blockers and Swift data race
- app.json: remove expo-build-properties plugin (not installed, caused
  prebuild to fail) and remove asset references to non-existent files
- LiveActivityModule.m: (int) -> (NSInteger) to match the 64-bit type
  the RCT bridge uses when coercing JS numbers
- LiveActivityModule.swift: nil currentActivity synchronously before
  the async end task to prevent startActivity from seeing a stale ref
  if called before the end task completes
- DoomscrollAttributes: replace unused startTime with a required
  sessionID string (ActivityAttributes must have at least one fixed field)
- SETUP.md: remove expo-build-properties references, add explicit
  verification checklist for testing the full system loop

+++ ios/LiveActivityModule/LiveActivityModule.m
@@ -4,14 +4,14 @@
 RCT_EXTERN_MODULE(LiveActivityModule, NSObject)
 
 RCT_EXTERN_METHOD(
-  startActivity:(int)elapsedSeconds
+  startActivity:(NSInteger)elapsedSeconds
   stage:(NSString *)stage
   resolve:(RCTPromiseResolveBlock)resolve
   reject:(RCTPromiseRejectBlock)reject
 )
 
 RCT_EXTERN_METHOD(
-  updateActivity:(int)elapsedSeconds
+  updateActivity:(NSInteger)elapsedSeconds
   stage:(NSString *)stage
   resolve:(RCTPromiseResolveBlock)resolve
   reject:(RCTPromiseRejectBlock)reject
